requirements.txt
streamlit
pymupdf
python-docx
openai
langchain
sentence-transformers
faiss-cpu
